requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
requests==2.31.0
